refactor : 반환값 인터페이스로 타입으로 수정
