+++ src/main/java/nextstep/ladder/domain/lines/point/Point.java
@@ -1,10 +1,8 @@
 package nextstep.ladder.domain.lines.point;
 
-import nextstep.ladder.domain.lines.point.impl.PointImpl;
-
 public interface Point {
 
-    PointImpl decideNextPoint(Boolean randomBoolean);
+    Point decideNextPoint(Boolean randomBoolean);
 
     boolean isPointFalse();
